requirements.txt
coremltools
typer
Pillow